quick fix : Redundant type cast

+++ src/main/java/tconstruct/blocks/traps/Punji.java
@@ -50,9 +50,9 @@ public class Punji extends MantleBlock {
             if (stack != null && stack.getItem() == Item.getItemFromBlock(TinkerWorld.punji)) {
                 world.setBlockMetadataWithNotify(x, y, z, meta + 1, 3);
                 world.playSoundEffect(
-                        (double) ((float) x + 0.5F),
-                        (double) ((float) y + 0.5F),
-                        (double) ((float) z + 0.5F),
+                        (float) x + 0.5F,
+                        (float) y + 0.5F,
+                        (float) z + 0.5F,
                         this.stepSound.soundName,
                         (this.stepSound.getVolume() + 1.0F) / 2.0F,
                         this.stepSound.getPitch() * 0.8F);
