+++ src/main/java/tconstruct/mechworks/landmine/behavior/BehaviorProjectile.java
@@ -47,9 +47,9 @@ public class BehaviorProjectile extends Behavior {
         }
 
         entity.setThrowableHeading(
-                (double) enumfacing.getFrontOffsetX(),
-                (double) ((float) enumfacing.getFrontOffsetY() + 0.1F),
-                (double) enumfacing.getFrontOffsetZ(),
+                enumfacing.getFrontOffsetX(),
+                (float) enumfacing.getFrontOffsetY() + 0.1F,
+                enumfacing.getFrontOffsetZ(),
                 1.1F,
                 6.0F);
         par1World.spawnEntityInWorld((Entity) entity);
