+++ src/main/java/tconstruct/world/itemblocks/WoolSlab1Item.java
@@ -45,9 +45,9 @@ public class WoolSlab1Item extends MultiItemBlock {
         if ((side == 1 && flag || side == 0 && !flag) && b == this.block && trueMeta == stack.getItemDamage()) {
             if (world.setBlock(x, y, z, Blocks.wool, trueMeta, 3)) {
                 world.playSoundEffect(
-                        (double) ((float) x + 0.5F),
-                        (double) ((float) y + 0.5F),
-                        (double) ((float) z + 0.5F),
+                        (float) x + 0.5F,
+                        (float) y + 0.5F,
+                        (float) z + 0.5F,
                         this.block.stepSound.soundName,
                         (this.block.stepSound.getVolume() + 1.0F) / 2.0F,
                         this.block.stepSound.getPitch() * 0.8F);
